requirements.txt
matplotlib
torch
torchaudio
torchvision
Flask
onnx
onnxruntime
Pillow
flask-cors